// GradeSession.Client/src/shared/api/generated/model/claimInviteResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * GradeSession.Server
 * OpenAPI spec version: 1.0
 */
import type { SessionParticipantRole } from './sessionParticipantRole';

export interface ClaimInviteResponse {
  /** @nullable */
  sessionId?: string | null;
  /** @nullable */
  committeeSlotId?: string | null;
  role?: SessionParticipantRole;
}